% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microorganisms.codes}
\alias{microorganisms.codes}
\title{Data Set with Common Microorganism Codes}
\format{
A \link{data.frame}:
\itemize{
\item \code{code}\cr Commonly used code of a microorganism
\item \code{mo}\cr ID of the microorganism in the \link{microorganisms} data set
}
}
\usage{
microorganisms.codes
}
\description{
A data set containing commonly used codes for microorganisms, from laboratory systems and WHONET.
}
\details{
The contents of this data set are extensively used by the \code{AMR} package, which allow standardised and reproducible antimicrobial resistance data analysis.
}
\seealso{
\code{\link[AMR:as.mo]{AMR::as.mo()}} \link{microorganisms}
}
\keyword{datasets}
