% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_package.R
\docType{package}
\name{microorganisms_pkg}
\alias{microorganisms}
\alias{microorganisms-package}
\title{The \code{microorganisms} Package}
\source{
To cite 'microorganisms' in publications use:

Berends MS, Luz CF, Friedrich AW, Sinha BNM, Albers CJ, Glasner C (2022). "AMR: An R Package for Working with Antimicrobial Resistance Data." \emph{Journal of Statistical Software}, \emph{104}(3), 1-31. \doi{10.18637/jss.v104.i03}.

A BibTeX entry for LaTeX users is:

\preformatted{
@Manual{,
  title = {microorganisms: Microbial Taxonomy Reference Data},
  author = {Matthijs S. Berends and Christian F. Luz and Dennis Souverein and Erwin E. A. Hassing},
  year = {2023},
  note = {https://msberends.github.io/microorganisms/, https://github.com/msberends/microorganisms},
}
}
}
\description{
Welcome to the \code{microorganisms} package.

The \code{microorganisms} package is a \href{https://msberends.github.io/microorganisms/#copyright}{free and open-source} R package with \href{https://en.wikipedia.org/wiki/Dependency_hell}{zero dependencies} to work with recent data about the microbial taxonomy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://msberends.github.io/microorganisms/}
  \item \url{https://github.com/msberends/microorganisms}
  \item Report bugs at \url{https://github.com/msberends/microorganisms/issues}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.s.berends@umcg.nl} (\href{https://orcid.org/0000-0001-7620-1800}{ORCID})

Authors:
\itemize{
  \item Christian F. Luz (\href{https://orcid.org/0000-0001-5809-5995}{ORCID}) [contributor]
  \item Dennis Souverein (\href{https://orcid.org/0000-0003-0455-0336}{ORCID}) [contributor]
  \item Erwin E. A. Hassing [contributor]
}

}
\keyword{internal}
