% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{package}
\alias{package}
\alias{_PACKAGE}
\title{The \code{microorganisms} Package}
\description{
Welcome to the \code{microorganisms} package.

The \code{microorganisms} package is a \href{https://msberends.github.io/microorganisms/#copyright}{free and open-source} R data package with \href{https://en.wikipedia.org/wiki/Dependency_hell}{zero dependencies} to work with recent data about the microbial taxonomy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://msberends.github.io/microorganisms/}
  \item \url{https://github.com/msberends/microorganisms}
  \item Report bugs at \url{https://github.com/msberends/microorganisms/issues}
}

}
\author{
\strong{Maintainer}: Matthijs S. Berends \email{m.s.berends@umcg.nl} (\href{https://orcid.org/0000-0001-7620-1800}{ORCID})

Authors:
\itemize{
  \item Christian F. Luz (\href{https://orcid.org/0000-0001-5809-5995}{ORCID}) [contributor]
  \item Dennis Souverein (\href{https://orcid.org/0000-0003-0455-0336}{ORCID}) [contributor]
  \item Erwin E. A. Hassing [contributor]
}

}
\keyword{internal}
