% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_data}
\alias{update_data}
\title{Update Microbial Taxonomic Data}
\usage{
update_data(repo = "https://github.com/msberends/microorganisms")
}
\arguments{
\item{repo}{a GitHub repository, which should contain '\verb{/data-raw/microorganisms.rds}', '\verb{/data-raw/microorganisms.codes.rds}' and '\verb{/data-raw/microorganisms.old.rds}'}
}
\description{
Retrieve the latest microbial taxonomy data from GitHub.
}
