% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microorganisms.codes}
\alias{microorganisms.codes}
\title{Data Set with 5 910 Common Microorganism Codes}
\format{
A \link[tibble:tibble]{tibble} with 5 910 observations and 2 variables:
\itemize{
\item \code{code}\cr Commonly used code of a microorganism
\item \code{mo}\cr ID of the microorganism in the \link{microorganisms} data set
}
}
\usage{
microorganisms.codes
}
\description{
A data set containing commonly used codes for microorganisms from laboratory systems.
}
\examples{
microorganisms.codes
}
\seealso{
\link{microorganisms}
}
\keyword{datasets}
